# openapi/openapi.yaml
components:
  schemas:
...
    handlers.ErrorResponse:
      properties:
        code:
          description: Code 是稳定的接口契约标识，一旦发布只增不改。
          example: APP_NOT_FOUND
          type: string
        message:
          description: Message 是面向前端展示的安全错误文案。
          example: 应用不存在
          type: string
      type: object
    handlers.InstanceResourceSamplesResponse:
      properties:
        samples:
          description: Samples 是按时间升序返回的实例资源采样或聚合桶。
          items:
            $ref: '#/components/schemas/service.InstanceResourceSampleResult'
          type: array
          uniqueItems: false
      type: object
...
    handlers.NodeInstancesResponse:
      properties:
        instances:
          description: Instances 是节点上承载的应用实例摘要。
          items:
            $ref: '#/components/schemas/service.NodeInstanceResult'
          type: array
          uniqueItems: false
      type: object
    handlers.NodeResourceSamplesResponse:
      properties:
        samples:
          description: Samples 是按时间升序返回的节点资源采样或聚合桶。
          items:
            $ref: '#/components/schemas/service.NodeResourceSampleResult'
          type: array
          uniqueItems: false
      type: object
...
    service.InstanceResourceSampleResult:
      description: CurrentResource 是该应用在节点上的最近一次实例资源采样。
      properties:
        container_status:
          description: ContainerStatus 是容器采样时状态；bucket 查询取桶内最新非空状态。
          type: string
        cpu_percent:
          description: CPUPercent 是实例 CPU 使用百分比；nil 表示该次采样缺失。
          type: number
        disk_read_bytes:
          description: DiskReadBytes 是实例磁盘读取累计字节数。
          type: integer
        disk_write_bytes:
          description: DiskWriteBytes 是实例磁盘写入累计字节数。
          type: integer
        last_error:
          description: LastError 是采样过程的最近错误，空字符串表示没有错误或无采样值。
          type: string
        memory_limit_bytes:
          description: MemoryLimitBytes 是实例内存限制字节数。
          type: integer
        memory_used_bytes:
          description: MemoryUsedBytes 是实例内存已用字节数。
          type: integer
        network_rx_bytes:
          description: NetworkRxBytes 是实例网络接收累计字节数。
          type: integer
        network_tx_bytes:
          description: NetworkTxBytes 是实例网络发送累计字节数。
          type: integer
        sampled_at:
          description: SampledAt 是采样或聚合桶时间，统一输出 UTC RFC3339。
          type: string
      type: object
...
    service.NodeInstanceResult:
      properties:
        app_id:
          description: AppID 是实例所属应用 ID。
          type: string
        container_id:
          description: ContainerID 是 runtime 侧容器 ID；未创建容器时为空。
          type: string
        current_resource:
          $ref: '#/components/schemas/service.InstanceResourceSampleResult'
        name:
          description: Name 是应用名称。
          type: string
        org_id:
          description: OrgID 是应用所属组织 ID，用于后续前端跳转和权限上下文。
          type: string
        owner_user_id:
          description: OwnerUserID 是应用所有者用户 ID。
          type: string
        runtime_node_id:
          description: RuntimeNodeID 是实例所在 runtime 节点 ID。
          type: string
        status:
          description: Status 是应用生命周期状态。
          type: string
      type: object
    service.NodeResourceSampleResult:
      properties:
        cpu_percent:
          description: CPUPercent 是节点 CPU 使用百分比；nil 表示该次采样缺失。
          type: number
        disk_total_bytes:
          description: DiskTotalBytes 是节点磁盘总字节数。
          type: integer
        disk_used_bytes:
          description: DiskUsedBytes 是节点磁盘已用字节数。
          type: integer
        instance_count:
          description: InstanceCount 是采样时节点承载的实例数量。
          type: integer
        last_error:
          description: LastError 是采样过程的最近错误，空字符串表示没有错误或无采样值。
          type: string
        memory_total_bytes:
          description: MemoryTotalBytes 是节点内存总字节数。
          type: integer
        memory_used_bytes:
          description: MemoryUsedBytes 是节点内存已用字节数。
          type: integer
        network_rx_bytes:
          description: NetworkRxBytes 是节点网络接收累计字节数。
          type: integer
        network_tx_bytes:
          description: NetworkTxBytes 是节点网络发送累计字节数。
          type: integer
        sampled_at:
          description: SampledAt 是采样或聚合桶时间，统一输出 UTC RFC3339。
          type: string
      type: object
...
  securitySchemes:
    BearerAuth:
      description: 登录后获得的 JWT access token，前缀 "Bearer "。
      in: header
      name: Authorization
      type: apiKey
...
info:
  description: 基于 Hermes Agent runtime 的多组织管理后台 API。
  title: Agent Runtime Manager API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /apps/{appId}/resources:
    get:
      description: 按应用 ID 查询实例资源趋势；权限沿用应用读取权限
      parameters:
      - description: 应用 ID
        in: path
        name: appId
        required: true
        schema:
          type: string
      - description: 开始时间 RFC3339，默认最近 7 天
        in: query
        name: from
        schema:
          type: string
      - description: 结束时间 RFC3339，默认当前时间
        in: query
        name: to
        schema:
          type: string
      - description: 聚合粒度，可选 5m / 1h；为空返回原始采样
        in: query
        name: bucket
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.InstanceResourceSamplesResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: 应用资源趋势
      tags:
      - apps
...
  /runtime-nodes/{nodeId}/instances:
    get:
      description: 平台管理员分页查看节点承载的应用实例及最近一次实例资源采样
      parameters:
      - description: 节点 ID
        in: path
        name: nodeId
        required: true
        schema:
          type: string
      - description: 每页条数（默认 50，最大 200）
        in: query
        name: limit
        schema:
          type: integer
      - description: 分页偏移（默认 0）
        in: query
        name: offset
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.NodeInstancesResponse'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: runtime 节点实例列表
      tags:
      - runtime-nodes
  /runtime-nodes/{nodeId}/instances/{appId}/resources:
    get:
      description: 平台管理员按节点和应用 ID 查询该实例资源趋势
      parameters:
      - description: 节点 ID
        in: path
        name: nodeId
        required: true
        schema:
          type: string
      - description: 应用 ID
        in: path
        name: appId
        required: true
        schema:
          type: string
      - description: 开始时间 RFC3339，默认最近 7 天
        in: query
        name: from
        schema:
          type: string
      - description: 结束时间 RFC3339，默认当前时间
        in: query
        name: to
        schema:
          type: string
      - description: 聚合粒度，可选 5m / 1h；为空返回原始采样
        in: query
        name: bucket
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.InstanceResourceSamplesResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: runtime 节点实例资源趋势
      tags:
      - runtime-nodes
  /runtime-nodes/{nodeId}/resources:
    get:
      description: 平台管理员按时间范围查询节点 CPU、内存、磁盘、网络和实例数量指标
      parameters:
      - description: 节点 ID
        in: path
        name: nodeId
        required: true
        schema:
          type: string
      - description: 开始时间 RFC3339，默认最近 7 天
        in: query
        name: from
        schema:
          type: string
      - description: 结束时间 RFC3339，默认当前时间
        in: query
        name: to
        schema:
          type: string
      - description: 聚合粒度，可选 5m / 1h；为空返回原始采样
        in: query
        name: bucket
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.NodeResourceSamplesResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: runtime 节点资源趋势
      tags:
      - runtime-nodes
...
servers:
- url: /api/v1